backend/ai_blog_app/api/urls.py
from django.urls import path, include
from .views import blog_details, blog_list
from drf_spectacular.views import (
    SpectacularAPIView,
    SpectacularSwaggerView,
    SpectacularRedocView,
)

urlpatterns = [
    path("blogs/", blog_list, name="blog-list"),
    path("blogs/<int:pk>/", blog_details, name="blog-detail"),
    # schema file
    path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
    # Swagger UI
    path(
        "api/docs/",
        SpectacularSwaggerView.as_view(url_name="schema"),
        name="swagger-ui",
    ),
    # Redoc UI
    path(
        "api/redoc/",
        SpectacularRedocView.as_view(url_name="schema"),
        name="redoc",
    ),
]
