backend/ai_blog_app/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
dotenv
psycopg2-binary
pytube
assemblyai
langchain
langchain-google-genai
nest_asyncio